Move MakotoValidatePlatformKeyAttestation to platformkeyattestation package
