+++ wallet-provider-service/src/main/kotlin/adapter/platformkeyattestation/MakotoValidatePlatformKeyAttestation.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.europa.ec.eudi.walletprovider.adapter.keyattestation
+package eu.europa.ec.eudi.walletprovider.adapter.platformkeyattestation
 
 import arrow.core.Either
 import arrow.core.raise.either
